Change anonymous function syntax
Anonymous functions now have the return type after a colon on the right
side of the argument list.

example:

what used to be

    (fn int (int x) => x * 2)

is now

    (fn (int x) : int => x * 2)

This is more in line with trends in language design. In particular, it
is almost identical to PHP anonymous function syntax, and very similar
to TypeScript syntax.